mydailyworktask3/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
requests>=2.28.0
datasets>=2.12.0
accelerate>=0.20.0